requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
idna==3.11
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
platformdirs==4.9.2
pycparser==3.0
pygments==2.19.2
python-dotenv==1.2.1
requests==2.32.5
rich==14.3.3
textual==8.0.0
typing-extensions==4.15.0
uc-micro-py==1.0.3
urllib3==2.6.3
